refactor: Remove custom hook useUpdateFavoriteOffer.
